+++ src/test/java/ru/netology/services/StatsServiceTest.java
@@ -9,9 +9,9 @@ class StatsServiceTest {
     @Test
     public void testSum() {
         StatsService statsService = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = statsService.sumSales(sales);
-        int expected = 8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = statsService.sumSales(sales);
+        long expected = 8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18;
 
         Assertions.assertEquals(expected, actual);
     }
@@ -19,18 +19,18 @@ class StatsServiceTest {
     @Test
     public void testAverage() {
         StatsService statsService = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = statsService.averageSales(sales);
-        int expected = (8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18) / 12;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = statsService.averageSales(sales);
+        long expected = (8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18) / 12;
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
     public void testMonthMaximum() {
         StatsService statsService = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = statsService.calcMonthMaxSale(sales);
-        int expected = 8;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = statsService.calcMonthMaxSale(sales);
+        long expected = 8;
         Assertions.assertEquals(expected, actual);
 
     }
@@ -38,27 +38,27 @@ class StatsServiceTest {
     @Test
     public void testMonthMinimum() {
         StatsService statsService = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = statsService.calcMonthMinSale(sales);
-        int expected = 9;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = statsService.calcMonthMinSale(sales);
+        long expected = 9;
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
     public void testSalesMonthsBelowAverage() {
         StatsService statsService = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = statsService.salesMonthsBelowAverage(sales);
-        int expected = 5;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = statsService.salesMonthsBelowAverage(sales);
+        long expected = 5;
         Assertions.assertEquals(expected, actual);
     }
 
     @Test
     public void testSalesMonthsAboveAverage() {
         StatsService statsService = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = statsService.salesMonthsAboveAverage(sales);
-        int expected = 5;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = statsService.salesMonthsAboveAverage(sales);
+        long expected = 5;
         Assertions.assertEquals(expected, actual);
     }
